services/image-harvest/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
httpx>=0.24.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
aiofiles>=0.8.0
minio>=7.1.0
redis>=4.3.4
pika>=1.3.0
tenacity>=8.0.1
prometheus-client>=0.14.1
streetlevel>=0.12.3
pillow>=10.0.0
pillow-heif>=0.15.0  # Primary HEIC decoder, works on all platforms
pyheif>=0.7.0        # Faster alternative for Linux/Mac
geocoder>=1.38.1
pytest>=7.0.0        # For testing
httpx>=0.24.0        # For async HTTP requests 